window.scroll() without params should be ignored or throw an error
The parameters are marked non-optional to make sure that a x,y parameter is specified to scroll. Modified the test case to consider this scenario. This is compatible with Firefox. IE reports no error and scrolls to page top, like chrome before the patch.

BUG=113962

Review URL: https://codereview.chromium.org/59863003

git-svn-id: svn://svn.chromium.org/blink/trunk@161798 bbb929c8-8fbe-4397-9dbb-9b2b20218538