Never look for the focused elment under nodes which are not in a document.
This makes http://jsperf.com/parse-html-type 10% faster.
78k ops/s vs. 70k ops/s on the "tags" test on my machine.

This is kinda a hack for this microbenchmark, but the change
is also simply correct.  Hayato is working on a more generic
fix to make EventPath faster (which we'll still hit
when the node is actually in a document and could have a
focused child).

I also fixed FullScreenElementStack's copy-pasta version
of this function to also check.  I suspect if we full-screened
and element on the page of this microbenchmark we'd find it
was another 10% slower due to FSES's version of this
"search for a possibly removed child" function.

R=esprehn
BUG=318251,318448

Review URL: https://codereview.chromium.org/66333018

git-svn-id: svn://svn.chromium.org/blink/trunk@161984 bbb929c8-8fbe-4397-9dbb-9b2b20218538