Move ownership of the TextResourceDecoder to DecodedDataDocumentParser
Continued from https://codereview.chromium.org/23455024/ 

This moves the Document's reference to the decoder to the parser. The
parser informs the document of changes in the encoding as well as two
bits indicating if the encoding was determined heuristically (to set up
the decoder properly for child frames) and if the decoder has encountered
any decoding errors (to generate errors for XML documents).

R=abarth@chromium.org, japhet@chromium.org
BUG=277886

Review URL: https://codereview.chromium.org/69823002

git-svn-id: svn://svn.chromium.org/blink/trunk@162079 bbb929c8-8fbe-4397-9dbb-9b2b20218538