Make free functions we always pass *this to into methods in ContainerNode
The code in ContainerNode.cpp was confusing since there's a bunch of
generic looking methods referencing newParent, but in reality we always
pass *this to them so they were only ever talking about the node that
was having appendChild etc. called on it.

This patch makes all those free functions into methods and removes some
wrapper functions that made the code harder to follow.

Review URL: https://codereview.chromium.org/73823003

git-svn-id: svn://svn.chromium.org/blink/trunk@162134 bbb929c8-8fbe-4397-9dbb-9b2b20218538