Add two more unit tests that check TextIterator's behavior on text fields.
This patch adds more tests in TextIteratorTest to check the behavior when
there is a text field in the iterating range. To reduce boilerplates in
the tests, I added a new utility function called testIteration().

There are a few interesting takeaways from the new tests. First, TextIterator
returns an empty string when it sees a text field. Second, if TextIterator is
called with the TextIteratorEntersTextControls option, it first returns
a newline when it sees a text field, and then it returns the value of
the text field next.

I'm not sure if this behavior is a bug. For now I'm just capturing the current
behavior without touching the implementation.

BUG=none
R=yosin@chromium.org

Review URL: https://codereview.chromium.org/72923004

git-svn-id: svn://svn.chromium.org/blink/trunk@162351 bbb929c8-8fbe-4397-9dbb-9b2b20218538